Merge: 전체 채용공고 조회 기능 추가 (#4)
* Rename: 엔티티 필드 명 오탈자 수정

* Feat: 전체 채용공고 조회 기능 추가

* Test: 전체 채용공고 조회 테스트 코드 추가

+++ src/main/java/com/wanted/model/JobPosting.java
@@ -28,7 +28,7 @@ public class JobPosting extends BaseEntity {
 
     private String imageUrl;
     private String position;
-    private int reward;
+    private Long reward;
     private String content;
 
     @ElementCollection
